requirements.txt
Flask==2.0.1
google-cloud-translate==3.0.2
deep-translator==1.5.5
openai==0.10.2
googletrans==4.0.0-rc1